backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
neo4j==5.28.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.6.1
torch==2.5.1
tqdm==4.67.1
python-dotenv==1.0.1
